==> code/01_data_beetles.Rmd <==
---
title: "NEON download"
author: "Kari Norman"
date: "10/15/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(neonUtilities)
library(lubridate)
```

Download data
```{r}
beetle_dir <- here::here("data", "beetles_raw.rda")
if (file.exists(beetle_dir)){
  load(beetle_dir)
} else{
  beetles_raw <- loadByProduct("DP1.10022.001", check.size = FALSE)
  usethis::use_data(beetles_raw)
}
```

Grab necessary columns from different tables. Important column info:
* sampleID: unique ID for each combination of domainID, siteID, plotID, trapID, and collectDate (unique ID for bet_fielddata table)
* subsampleID: unique ID for each group of individuals, multiple per sample (unique ID for bet_sorting); unclear why there are muliple entries for the same classification
* individualID: unique ID for each individual identified by taxonomists (unique ID for bet_parataxonimist table)

There are multiple levels of taxonomic identification, need to decide which level of rigor we care about (NEON id often just to family, taxonomist ID to species but only for a subset, DNA ID for even smaller subset).

This first chunk also adds a new `trappingDays` column to account for effort. For each unique siteID/plotID/trapID/eventID combination (so each sample event for each trap) the number of days the trap was set before sampling is calculated.

There should be one boutID for each collectDate at each site. eventID doesn't quite do the trick since it can be associated with multiple sample days (usually consecutive) and there can be more than one eventID for a collectDate within a site (some plots have different eventID). So we create a new boutID that identifies a sampling period on a site.
```{r}
#function for finding the statistical mode
Mode <- function(x) {
  ux <- unique(x)
  ux[which.max(tabulate(match(x, ux)))]
}

data <- beetles_raw$bet_fielddata %>%
  filter(sampleCollected == "Y") %>% #there's an entry for every trap, whether or not they got samples, only want ones with samples
  select(sampleID, domainID, siteID, plotID, trapID, setDate, collectDate, eventID, trappingDays) %>%
  #eventID's are inconsistently separated by periods, so remove
  mutate(eventID = str_remove_all(eventID, "[.]")) %>%
  mutate(trappingDays = interval(ymd(setDate), ymd(collectDate)) %/% days(1))

#Find the traps that have multiple collectDates/bouts for the same setDate
#need to calculate their trap days from the previous collectDate, not the setDate
adjTrappingDays <- data %>%
  select(plotID, trapID, setDate, collectDate, trappingDays, eventID) %>%
  group_by_at(vars(-collectDate, -trappingDays, -eventID)) %>%
  filter(n_distinct(collectDate) > 1) %>%
  group_by(plotID, trapID, setDate) %>%
  mutate(diffTrappingDays = trappingDays - min(trappingDays)) %>%
  mutate(adjTrappingDays = case_when(
    diffTrappingDays == 0 ~ trappingDays,
    TRUE ~ diffTrappingDays
  )) %>%
  select(-c(trappingDays, diffTrappingDays))

data <- data %>%
  #update with adjusted trapping days where needed
  left_join(adjTrappingDays) %>%
  mutate(trappingDays = case_when(
    !is.na(adjTrappingDays) ~ adjTrappingDays,
    TRUE ~ trappingDays
  )) %>%
  select(-adjTrappingDays, -setDate) %>%
  #for some eventID's (bouts) collection happened over two days,
  #change collectDate to the date that majority of traps were collected on
  group_by(eventID) %>%
  mutate(collectDate = Mode(collectDate)) %>%
  ungroup() %>%
  #there are also some sites for which all traps were set and collect on the same date, but have multiple eventID's
  #we want to consider that as all one bout so we'll just create a new ID based on the site and collectDate
  unite(boutID, siteID, collectDate, remove = FALSE) %>%
  select(-eventID) %>%
  #and join to sample data
  left_join(beetles_raw$bet_sorting %>%
              filter(sampleType %in% c("carabid", "other carabid")) %>% #only want carabid samples, not bycatch
              select(sampleID, subsampleID, sampleType, taxonID, scientificName, taxonRank, individualCount,identificationQualifier),
            by = "sampleID") %>%
  filter(!is.na(subsampleID)) #even though they were marked a sampled, some collection times don't acutally have any samples
```

Join taxonomic data from pinning with the sorting data
```{r}
# Replace sorting taxon info with pinning taxon info (people that pin specimens are more experienced with taxonomy), where available
data_pin <- data %>%
  left_join(beetles_raw$bet_parataxonomistID %>% select(subsampleID, individualID, taxonID, scientificName, taxonRank,identificationQualifier), by = "subsampleID") %>%
  mutate_if(is.factor, as.character) %>%
  mutate(taxonID = ifelse(is.na(taxonID.y), taxonID.x, taxonID.y)) %>%
  mutate(taxonRank = ifelse(is.na(taxonRank.y), taxonRank.x, taxonRank.y)) %>%
  mutate(scientificName = ifelse(is.na(scientificName.y), scientificName.x, scientificName.y)) %>%
  mutate(identificationSource = ifelse(is.na(scientificName.y), "sort", "pin")) %>%
  mutate (identificationQualifier = ifelse(is.na(taxonID.y), identificationQualifier.x, identificationQualifier.y)) %>%
  select(-ends_with(".x"), -ends_with(".y"))

#some subsamples weren't fully ID'd by the pinners, so we have to recover the unpinned-individuals
lost_indv <- data_pin %>% 
  filter(!is.na(individualID)) %>%
  group_by(subsampleID, individualCount) %>%
  summarise(n_ided = n_distinct(individualID)) %>% 
  filter(n_ided < individualCount) %>%
  mutate(unidentifiedCount = individualCount - n_ided) %>%
  select(subsampleID, individualCount = unidentifiedCount) %>%
  left_join(data %>% select(-individualCount), by = "subsampleID") %>%
  mutate(identificationSource = "sort")

#add unpinned-individuals back to the pinned id's, adjust the individual counts so pinned individuals have a count of 1
data_pin <- data_pin %>%
  mutate(individualCount = ifelse(identificationSource == "sort", individualCount, 1)) %>%
  bind_rows(lost_indv)
```

Join expert data to existing pinning and sorting data
```{r}
#There are ~10 individualID's for which experts ID'd more than one species (not all experts agreed), we want to exclude those expert ID's as per Katie Levan's suggestion
ex_expert_id <- beetles_raw$bet_expertTaxonomistIDProcessed %>% 
  group_by(individualID) %>% 
  filter(n_distinct(taxonID) > 1) %>% 
  pull(individualID)

# Add expert taxonomy info, where available
data_expert <- left_join(data_pin, 
                      select(beetles_raw$bet_expertTaxonomistIDProcessed,
                             individualID,taxonID,scientificName,taxonRank,identificationQualifier) %>%
                        filter(!individualID %in% ex_expert_id), #exclude ID's that have unresolved expert taxonomy
                      by = 'individualID', na_matches = "never") %>% distinct()

# Replacement old taxon info with expert info, where available
# NOTE - This is repetitive with the code snippet above, and if you want to do it this way you can just combine the calls into one chunk. BUT, you may
#     want to do more than this, as it is *just* a replacement of IDs for individual beetles that an expert identified. If the expert identified
#           a sample as COLSP6 instead of CARSP14, though, then all CARSP14 from that trap on that date should probably be updated to COLSP6…
data_expert <- data_expert %>%
  mutate_if(is.factor, as.character) %>%
  mutate(taxonID = ifelse(is.na(taxonID.y), taxonID.x, taxonID.y)) %>%
  mutate(taxonRank = ifelse(is.na(taxonRank.y), taxonRank.x, taxonRank.y)) %>%
  mutate(scientificName = ifelse(is.na(scientificName.y), scientificName.x, scientificName.y)) %>%
  mutate(identificationSource = ifelse(is.na(scientificName.y), identificationSource, "expert")) %>%
  mutate (identificationQualifier = ifelse(is.na(taxonID.y), identificationQualifier.x, identificationQualifier.y)) %>%
  select(-ends_with(".x"), -ends_with(".y"))

data_expert
```

Get raw counts table
```{r}
beetles_counts <- data_expert %>%
  select(-c(subsampleID, sampleType, individualID, identificationSource, identificationQualifier)) %>%
  group_by_at(vars(-individualCount)) %>%
  summarise(count = sum(individualCount)) %>%
  ungroup()

write_csv(beetles_counts, "data/beetles_counts.csv")
```
